Complete overhaul using tabs
I was wrong when I said it was "built-in" functionality, but I went
ahead and implemented a tab layout so that slides can be easily
navigated. I would like to eventually integrate the graph into the
tab-selector, so that the user can click on the part of the graph that
they want to change. but before doing anything like that, we need to
have another talk about layout, and I'd like to see the data structures
you're working with on the back-end so I can first and foremost ensure
that I'm collecting the appropriate data from the user.